+++ NEWS.md
@@ -1,3 +1,7 @@
+# spant 2.25.0
+* scale_amp_molar function has been depreciated. Please use scale_amp_legacy
+instead.
+
 # spant 2.24.0
 * Added argument to adjust the legend font size in ortho3 function.
 * Improved metabolite simulation performance thanks to the expm package recently